requirements.txt
pillow
pytesseract
discord.py
